+++ include/wx/odcombo.h
@@ -308,7 +308,7 @@ public:
     }
 
     // wxControlWithItems methods
-    unsigned int GetCount() const override;
+    size_t GetCount() const override;
     std::string GetString(unsigned int n) const override;
     void SetString(unsigned int n, const std::string& s) override;
     int FindString(std::string_view s, bool bCase = false) const override;
